requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
asyncpg
pydantic
ollama
pyjwt
pytest
pytest-asyncio
httpx
